+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - `context-controls-no-hover`: New feature. Prevents the context controls from displaying description while hovering over
+- `cxos-leftalign-delete-filled`: New feature. Moves 'Delete Filled' button to the left in CXOS
 - `prod-hide-percent`: New feature. Hides percent value from production lines.
 
 ### Fixed